refactor(server): extract serve bind gates
Move ensure_bind_allowed, ensure_metrics_bind_allowed, and
is_loopback_host into serve/bind.rs. Public CLI flags and
boot path are unchanged.
